test: reorganization of the 'private_data' unit test
