+++ tests/unit/CMakeLists.txt
@@ -12,6 +12,6 @@ add_subdirectory(info-test)
 add_subdirectory(log)
 add_subdirectory(mr)
 add_subdirectory(peer)
-add_subdirectory(private_data-test)
+add_subdirectory(private_data)
 add_subdirectory(template-test)
 add_subdirectory(utils-test)
