--- tests/unit/private_data-test/CMakeLists.txt
@@ -1,18 +1,0 @@
-#
-# SPDX-License-Identifier: BSD-3-Clause
-# Copyright 2020, Intel Corporation
-#
-
-include(../../cmake/ctest_helpers.cmake)
-
-build_test_src(UNIT NAME private_data-test SRCS
-       private_data-test.c
-       ${LIBRPMA_SOURCE_DIR}/private_data.c)
-
-target_compile_definitions(private_data-test PRIVATE TEST_MOCK_ALLOC)
-
-set_target_properties(private_data-test
-       PROPERTIES
-       LINK_FLAGS "-Wl,--wrap=_test_malloc")
-
-add_test_generic(NAME private_data-test TRACERS none)
